+++ example/yeedart_example.dart
@@ -3,16 +3,22 @@ import 'dart:io';
 import 'package:yeedart/yeedart.dart';
 
 Future<void> main() async {
+  // Discover devices
 //  final responses = await Yeelight.discover();
 //  final response = responses.first;
-//
-//  final device = YeeDevice(address: response.address, port: response.port);
+//  final device = Device(address: response.address, port: response.port);
+//  device.turnOn();
+//  await Future<void>.delayed(const Duration(seconds: 3));
+//  device.turnOff();
+//  device.disconnect();
 
+  // Connect directly to device
   final device = Device(
     address: InternetAddress("192.168.1.183"),
     port: 55443,
   );
 
+  // ignore: avoid_print
   print(await device.getProps(id: 1, parameters: [
     "name",
     "model",
@@ -26,9 +32,8 @@ Future<void> main() async {
     "sat",
   ]));
 
-  //await device.turnOn();
-  //await device.turnOff();
-  //await device.setScene(scene: Scene.color(color: Colors.red, brightness: 20));
-  //await device.stopFlow();
+  await device.turnOn();
+  await Future<void>.delayed(const Duration(seconds: 3));
+  await device.turnOff();
   device.disconnect();
 }