+++ lib/src/command/command_sender.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:yeedart/src/command/command.dart';
 import 'package:yeedart/src/response/command_response.dart';
 
@@ -6,6 +8,8 @@ import 'package:yeedart/src/response/command_response.dart';
 /// To create custom command sender, implement this class.
 /// Default implementation is [TCPCommandSender].
 abstract class CommandSender {
+  Stream<Uint8List> get connectionStream;
+
   /// Sends command to the device.
   Future<CommandResponse> sendCommand(Command command);