+++ lib/src/response/command_response.dart
@@ -11,6 +11,8 @@ class CommandResponse {
   /// If command fails, error will be returned.
   final Map<String, dynamic> error;
 
+  CommandResponse({this.id, this.result, this.error});
+
   /// Creates [CommandResponse] from parsed JSON.
   CommandResponse.fromJson(Map<String, dynamic> parsed)
       : id = parsed["id"] as int,
@@ -19,6 +21,9 @@ class CommandResponse {
 
   /// Indicates whether command was successfully executed or not.
   bool get hasError => error != null;
+
+  /// Returns raw response (as string).
+  String get raw => _toJson();
 
   String _toJson() {
     return hasError
@@ -39,6 +44,7 @@ class CommandResponse {
   bool operator ==(Object other) {
     return identical(this, other) ||
         other is CommandResponse &&
+            runtimeType == other.runtimeType &&
             id == other.id &&
             result.toString() == other.result.toString();
   }